fix: rename resizedImages
