+++ src/controllers/resize.ts
@@ -3,7 +3,7 @@ import { Controller, Ctx, Get, Params, Query } from 'amala'
 import { IPFS_TIMEOUT } from '../constants'
 import { Readable } from 'stream'
 import { badRequest, internal, notFound } from '@hapi/boom'
-import { getCachedResize, setCachedResize } from '../helpers/resizedImages'
+import { getCachedResize, setCachedResize } from '../helpers/resizedImageCache'
 import { getContentType } from '../helpers/contentTypes'
 import ipfs, { getDataFromIPFS } from '../helpers/ipfs'
 import sharp from 'sharp'
